test(HYPRE): Switch FlexGMRES preconditioner from SMG to PFMG for periodic tests
Changes the preconditioner used with the FlexGMRES solver in
`EffectiveDiffusivityHypre::solve()` from SMG to PFMG.

This change is motivated by HYPRE documentation indicating that:
- The SMG solver has a strict requirement for grid sizes in periodic
  dimensions to be a power-of-two.
- The PFMG solver has a less strict recommendation for periodic dimensions
  (preferring sizes close to a power-of-two for performance but not
  necessarily a hard failure condition).

The previous "Generic error" (Error Code 1) at `HYPRE_StructMatrixCreate`
when using a 100x100x100 periodic domain is suspected to be related to
these non-power-of-two dimensions, potentially due to internal checks
anticipating the use of SMG or similar strict solvers.

By switching to PFMG as the preconditioner, this commit attempts to
determine if the HYPRE setup can proceed further with non-power-of-two
periodic dimensions. If `HYPRE_StructMatrixCreate` still fails, it
would indicate a more fundamental issue with creating matrices on
such periodic grids in HYPRE. If it proceeds, it suggests the strict
power-of-two limitation was tied more directly to the SMG setup.